Move all components into JaxGym
