+++ src/jaxgym/ode.py
@@ -1,9 +1,11 @@
 import jax
 import jax.numpy as jnp
+import jax_dataclasses as jdc
 from functools import partial
 import diffrax
 import numpy as np
 import tqdm.auto as tqdm
+from .ray import Ray
 
 
 @partial(jax.jit, static_argnums=(0,))
@@ -230,3 +232,27 @@ def electron_equation_of_motion_DA(z, x, args):
     dopl = (u / u0) ** (1 / 2) * (v) ** (1 / 2)
 
     return np.array([dx, dy, ddx, ddy, dopl])
+
+
+@jdc.pytree_dataclass
+class ODE:
+    z: float
+    z_end: float
+    phi_lambda: callable
+    E_lambda: callable
+
+    def step(self, ray: Ray) -> Ray:
+        in_state = jnp.array([ray.x, ray.y, ray.dx, ray.dy, ray.pathlength])
+
+        z_start = self.z
+        z_end = self.z_end
+
+        u0 = self.phi_lambda(0.0, 0.0, z_start).astype(jnp.float64)
+
+        out_state, out_z = solve_ode(
+            in_state, z_start, z_end, self.phi_lambda, self.E_lambda, u0
+        )
+
+        x, y, dx, dy, opl = out_state
+
+        return Ray(x=x, y=y, dx=dx, dy=dy, _one=ray._one, pathlength=opl, z=out_z)
